__init.py__, soundfigure into wfs handling
