+++ sfs/mono/__init__.py
@@ -9,19 +9,37 @@
     wfs
 
     source
-    soundfigure
 
 """
+from . import source as _source
+
+
+def secondary_source_point(omega, c):
+    """Create a point source for use in `sfs.mono.synthesize()`."""
+
+    def secondary_source(position, _, grid):
+        return _source.point(omega, position, grid, c)
+
+    return secondary_source
+
+
+def secondary_source_line(omega, c):
+    """Create a line source for use in `sfs.mono.synthesize()`."""
+
+    def secondary_source(position, _, grid):
+        return _source.line(omega, position, grid, c)
+
+    return secondary_source
+
+
 import numpy as _np
 
+from . import source
 
 from . import esaedge
 from . import nfchoa
 from . import sdm
 from . import wfs
-
-from . import source
-from . import soundfigure
 
 from .. import array as _array
