+++ sfs/mono/nfchoa.py
@@ -37,7 +37,7 @@ from numpy.core.umath_tests import inner1d  # element-wise inner product
 from scipy.special import jn, hankel2
 from .. import util
 from .. import default
-from . import source as _source
+from . import secondary_source_point
 
 
 def plane_2d(omega, x0, r0, n=[0, 1, 0], max_order=None, c=None):
@@ -239,21 +239,3 @@ def plane_25d(omega, x0, r0, n=[0, 1, 0], max_order=None, c=None):
         d += (-1j)**abs(m) / (k * hn2[abs(m)]) * np.exp(1j * m * (phi0 - phi))
     selection = util.source_selection_all(len(x0))
     return 2*1j / r0 * d, selection, secondary_source_point(omega, c)
-
-
-def secondary_source_point(omega, c):
-    """Create a point source for use in `sfs.mono.synthesize()`."""
-
-    def secondary_source(position, _, grid):
-        return _source.point(omega, position, grid, c)
-
-    return secondary_source
-
-
-def secondary_source_line(omega, c):
-    """Create a line source for use in `sfs.mono.synthesize()`."""
-
-    def secondary_source(position, _, grid):
-        return _source.line(omega, position, grid, c)
-
-    return secondary_source
